requirements.txt
asgiref==3.8.1
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.7
pillow==10.4.0
PyJWT==1.7.1
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
